backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
pydantic
python-dotenv
pydantic-settings
pytz
redis
python-socketio
slowapi